Corrected doctest display bug
If specialized, StringMaker<T> needs to be included in every file
that uses T in its CHECK, REQUIRE... If not, apparently there is
a linker problem with a competition between doctest defaults
(e.g. (?)==(?) ) and the user-defined StringMaker<T>.

To circumvent the problem, the idea is that everyone #includes std_e/unit_test/doctest.hpp.